+++ resources/lang/en.json
@@ -326,25 +326,6 @@
     "view_options": "View Options",
     "toggle_view": "Toggle View",
     "toggle_view_desc": "Alternate view (terrain/countries)",
-    "build_controls": "Build Controls",
-    "build_city": "Build City",
-    "build_city_desc": "Build a City under your cursor.",
-    "build_factory": "Build Factory",
-    "build_factory_desc": "Build a Factory under your cursor.",
-    "build_defense_post": "Build Defense Post",
-    "build_defense_post_desc": "Build a Defense Post under your cursor.",
-    "build_port": "Build Port",
-    "build_port_desc": "Build a Port under your cursor.",
-    "build_warship": "Build Warship",
-    "build_warship_desc": "Build a Warship under your cursor.",
-    "build_missile_silo": "Build Missile Silo",
-    "build_missile_silo_desc": "Build a Missile Silo under your cursor.",
-    "build_sam_launcher": "Build SAM Launcher",
-    "build_sam_launcher_desc": "Build a SAM Launcher under your cursor.",
-    "build_atom_bomb": "Build Atom Bomb",
-    "build_atom_bomb_desc": "Build an Atom Bomb under your cursor.",
-    "build_hydrogen_bomb": "Build Hydrogen Bomb",
-    "build_hydrogen_bomb_desc": "Build a Hydrogen Bomb under your cursor.",
     "attack_ratio_controls": "Attack Ratio Controls",
     "attack_ratio_up": "Increase Attack Ratio",
     "attack_ratio_up_desc": "Increase attack ratio by 10%",
