+++ resources/maps/africa/manifest.json
@@ -103,7 +103,7 @@
     {
       "coordinates": [1094, 1093],
       "flag": "cd",
-      "name": "DR Congo",
+      "name": "Democratic Republic of the Congo",
       "strength": 2
     },
     {
