+++ resources/maps/europe/manifest.json
@@ -169,7 +169,7 @@
     {
       "coordinates": [517, 1483],
       "flag": "dz",
-      "name": "Algeria",
+      "name": "People's Democratic Republic of Algeria",
       "strength": 1
     },
     {
