+++ resources/maps/gatewaytotheatlantic/manifest.json
@@ -85,7 +85,7 @@
     {
       "coordinates": [1609, 1837],
       "flag": "dz",
-      "name": "Algeria",
+      "name": "People's Republic of Algeria'",
       "strength": 2
     },
     {
