+++ resources/maps/oceania/manifest.json
@@ -85,7 +85,7 @@
     {
       "coordinates": [143, 37],
       "flag": "la",
-      "name": "Lao PDR",
+      "name": "Lao People's Democratic Republic",
       "strength": 1
     },
     {
@@ -115,7 +115,7 @@
     {
       "coordinates": [834, 215],
       "flag": "fm",
-      "name": "Micronesia",
+      "name": "Federated States of Micronesia",
       "strength": 1
     },
     {
