+++ resources/maps/world/manifest.json
@@ -253,7 +253,7 @@
     {
       "coordinates": [1074, 508],
       "flag": "cd",
-      "name": "DR Congo",
+      "name": "The Democratic Republic of the Congo",
       "strength": 1
     },
     {
