+++ src/client/InputHandler.ts
@@ -2,7 +2,6 @@ import { EventBus, GameEvent } from "../core/EventBus";
 import { UnitType } from "../core/game/Game";
 import { UnitView } from "../core/game/GameView";
 import { UserSettings } from "../core/game/UserSettings";
-import { UIState } from "./graphics/UIState";
 import { ReplaySpeedMultiplier } from "./utilities/ReplaySpeedMultiplier";
 
 export class MouseUpEvent implements GameEvent {
@@ -76,7 +75,7 @@ export class RefreshGraphicsEvent implements GameEvent {}
 export class TogglePerformanceOverlayEvent implements GameEvent {}
 
 export class ToggleStructureEvent implements GameEvent {
-  constructor(public readonly structureTypes: UnitType[] | null) {}
+  constructor(public readonly structureType: UnitType | null) {}
 }
 
 export class ShowBuildMenuEvent implements GameEvent {
@@ -137,36 +136,14 @@ export class InputHandler {
   private readonly PAN_SPEED = 5;
   private readonly ZOOM_SPEED = 10;
 
-  private readonly userSettings: UserSettings = new UserSettings();
+  private userSettings: UserSettings = new UserSettings();
 
   constructor(
-    public uiState: UIState,
     private canvas: HTMLCanvasElement,
     private eventBus: EventBus,
   ) {}
 
   initialize() {
-    let saved: Record<string, string> = {};
-    try {
-      const parsed = JSON.parse(
-        localStorage.getItem("settings.keybinds") ?? "{}",
-      );
-      // flatten { key: {key, value} } → { key: value } and accept legacy string values
-      saved = Object.fromEntries(
-        Object.entries(parsed)
-          .map(([k, v]) => {
-            if (v && typeof v === "object" && "value" in (v as any)) {
-              return [k, (v as any).value as string];
-            }
-            if (typeof v === "string") return [k, v];
-            return [k, undefined];
-          })
-          .filter(([, v]) => typeof v === "string" && v !== "Null"),
-      ) as Record<string, string>;
-    } catch (e) {
-      console.warn("Invalid keybinds JSON:", e);
-    }
-
     this.keybinds = {
       toggleView: "Space",
       centerCamera: "KeyC",
@@ -176,22 +153,13 @@ export class InputHandler {
       moveRight: "KeyD",
       zoomOut: "KeyQ",
       zoomIn: "KeyE",
-      attackRatioDown: "KeyT",
-      attackRatioUp: "KeyY",
+      attackRatioDown: "Digit1",
+      attackRatioUp: "Digit2",
       boatAttack: "KeyB",
       groundAttack: "KeyG",
       modifierKey: "ControlLeft",
       altKey: "AltLeft",
-      buildCity: "Digit1",
-      buildFactory: "Digit2",
-      buildPort: "Digit3",
-      buildDefensePost: "Digit4",
-      buildMissileSilo: "Digit5",
-      buildSamLauncher: "Digit6",
-      buildAtomBomb: "Digit7",
-      buildHydrogenBomb: "Digit8",
-      buildWarship: "Digit9",
-      ...saved,
+      ...JSON.parse(localStorage.getItem("settings.keybinds") ?? "{}"),
     };
 
     // Mac users might have different keybinds
@@ -298,7 +266,6 @@ export class InputHandler {
       if (e.code === "Escape") {
         e.preventDefault();
         this.eventBus.emit(new CloseViewEvent());
-        this.uiState.ghostStructure = null;
       }
 
       if (
@@ -362,51 +329,6 @@ export class InputHandler {
       if (e.code === this.keybinds.centerCamera) {
         e.preventDefault();
         this.eventBus.emit(new CenterCameraEvent());
-      }
-
-      if (e.code === this.keybinds.buildCity) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.City;
-      }
-
-      if (e.code === this.keybinds.buildFactory) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.Factory;
-      }
-
-      if (e.code === this.keybinds.buildPort) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.Port;
-      }
-
-      if (e.code === this.keybinds.buildDefensePost) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.DefensePost;
-      }
-
-      if (e.code === this.keybinds.buildMissileSilo) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.MissileSilo;
-      }
-
-      if (e.code === this.keybinds.buildSamLauncher) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.SAMLauncher;
-      }
-
-      if (e.code === this.keybinds.buildAtomBomb) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.AtomBomb;
-      }
-
-      if (e.code === this.keybinds.buildHydrogenBomb) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.HydrogenBomb;
-      }
-
-      if (e.code === this.keybinds.buildWarship) {
-        e.preventDefault();
-        this.uiState.ghostStructure = UnitType.Warship;
       }
 
       // Shift-D to toggle performance overlay
@@ -567,10 +489,6 @@ export class InputHandler {
 
   private onContextMenu(event: MouseEvent) {
     event.preventDefault();
-    if (this.uiState.ghostStructure !== null) {
-      this.uiState.ghostStructure = null;
-      return;
-    }
     this.eventBus.emit(new ContextMenuEvent(event.clientX, event.clientY));
   }
