+++ src/client/graphics/UIState.ts
@@ -1,6 +1,3 @@
-import { UnitType } from "../../core/game/Game";
-
 export interface UIState {
   attackRatio: number;
-  ghostStructure: UnitType | null;
 }
