+++ src/client/graphics/layers/RailroadLayer.ts
@@ -9,7 +9,6 @@ import {
   RailType,
 } from "../../../core/game/GameUpdates";
 import { GameView } from "../../../core/game/GameView";
-import { TransformHandler } from "../TransformHandler";
 import { Layer } from "./Layer";
 import { getBridgeRects, getRailroadRects } from "./RailroadSprites";
 
@@ -28,10 +27,7 @@ export class RailroadLayer implements Layer {
   private nextRailIndexToCheck = 0;
   private railTileList: TileRef[] = [];
 
-  constructor(
-    private game: GameView,
-    private transformHandler: TransformHandler,
-  ) {
+  constructor(private game: GameView) {
     this.theme = game.config().theme();
   }
 
@@ -95,11 +91,6 @@ export class RailroadLayer implements Layer {
 
   renderLayer(context: CanvasRenderingContext2D) {
     this.updateRailColors();
-    if (this.transformHandler.scale <= 2) {
-      // When zoomed out, don't show the railroads
-      // to prevent map clutter.
-      return;
-    }
     context.drawImage(
       this.canvas,
       -this.game.width() / 2,
