+++ src/core/game/GameView.ts
@@ -264,11 +264,11 @@ export class PlayerView {
       : this._defendedBorderColors.dark;
   }
 
-  async actions(tile?: TileRef): Promise<PlayerActions> {
+  async actions(tile: TileRef): Promise<PlayerActions> {
     return this.game.worker.playerInteraction(
       this.id(),
-      tile && this.game.x(tile),
-      tile && this.game.y(tile),
+      this.game.x(tile),
+      this.game.y(tile),
     );
   }
